requirements.txt
aiosignal==1.3.1
anyio==4.3.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
Brlapi==0.7.0
cachetools==5.3.3
certifi==2019.11.28
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.3
comm==0.2.2
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.16
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
descartes==1.1.0
distro==1.4.0
distro-info===0.23ubuntu1
dm-tree==0.1.8
entrypoints==0.3
exceptiongroup==1.2.0
executing==2.0.1
Farama-Notifications==0.0.4
fastjsonschema==2.19.1
filelock==3.13.1
fire==0.6.0
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gymnasium==0.28.1
h11==0.14.0
httpcore==1.0.4
httplib2==0.14.0
httpx==0.27.0
idna==2.8
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.3
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
jax-jumpy==1.0.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
keyring==18.0.1
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy_loader==0.3
louis==3.12.0
lz4==4.3.3
macaroonbakery==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.8
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
netifaces==0.10.4
networkx==3.1
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.24.4
nuscenes-devkit==1.1.11
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
opencv-python==4.9.0.80
overrides==7.7.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==5.26.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pycairo==1.16.2
pycocotools==2.0.7
pycparser==2.21
pycups==1.9.73
pygame==2.1.3
Pygments==2.17.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.1.2
pyquaternion==0.9.9
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian===0.1.36ubuntu1
python-json-logger==2.0.7
pytz==2024.1
PyWavelets==1.4.1
pyxdg==0.26
PyYAML==5.4
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
ray==2.10.0
referencing==0.34.0
reportlab==3.5.34
requests==2.31.0
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
SecretStorage==2.3.1
Send2Trash==1.8.2
Shapely==1.8.5.post1
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
systemd-python==234
tensorboardX==2.6.2.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.3.0
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
triton==2.2.0
typer==0.10.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.10
xkit==0.0.0
zipp==3.18.1
